refractor: move /about to (view) route group
